% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-accumulate.outcomes.combination.R
\name{accum.outcomes.combination}
\alias{accum.outcomes.combination}
\title{Interventions}
\usage{
accum.outcomes.combination(
  input.parameters,
  pop.in,
  current.int,
  sums.only = TRUE
)
}
\arguments{
\item{input.parameters}{model inputs}

\item{pop.in}{populations in each compartment over all evaluation time}

\item{current.int}{selected intervention}

\item{sums.only}{logical to only output the total of results for full time period (default = TRUE)}
}
\value{
List with total and period-stratified outcomes (disaggregated costs, QALYs)
}
\description{
\code{accum.outcomes.combination} Costs & QALYs calculations for each model period
}
