% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-comb.eva.func.R
\name{comb.eva.func}
\alias{comb.eva.func}
\title{Interventions}
\usage{
comb.eva.func(input.parameters, current.int)
}
\arguments{
\item{input.parameters}{model inputs}

\item{current.int}{selected intervention}
}
\value{
Vector of incidence and cost-effectiveness outcomes
}
\description{
\code{comb.eva.func} combination evaluation function: generate population list, incidence results, CEA results
}
