% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CascadeCEA-Model-0-Function-FoI.R
\name{FoI}
\alias{FoI}
\title{Force of infection equations}
\usage{
FoI(
  y,
  no,
  uoC,
  ns,
  usC,
  eO,
  eS,
  sigmaFM,
  sigmaMF,
  sigmaM,
  tau,
  eff.prep,
  d,
  s,
  eff.oat,
  cov.ssp,
  eff.ssp = eff.ssp,
  s.multi,
  bal = F
)
}
\arguments{
\item{y}{number of individuals in each compartment (n.gp*19 matrix)}

\item{no}{number of opposite sexual partners for each compartment (n.gp*19 matrix)}

\item{uoC}{condom use adjustment term for opposite sex (n.gp*19 matrix)}

\item{ns}{number of same sexual partners for each compartment (18*19 matrix)}

\item{usC}{condom use adjustment term for same sex (18*19 matrix)}

\item{eO}{assortative coefficient for heterosexual mixing between ethnic groups (vector of 3 for white, black, hispanic)}

\item{eS}{assortative coefficient for homosexual mixing between ethnic groups   (vector of 3 for white, black, hispanic)}

\item{sigmaFM}{probability of transmission by F->M for 16 HIV+ states}

\item{sigmaMF}{probability of transmission by M->F for 16 HIV+ states}

\item{tau}{probability of transmission by injection for 16 HIV+ states}

\item{eff.prep}{\% reduction in risk of infection for PrEP}

\item{d}{number of injection}

\item{s}{proportion of injections that are shared (9 groups)}

\item{eff.oat}{\% reduction in # of shared injections reduced due to OAT}

\item{cov.ssp}{coverage of syringe services program}

\item{simgaM}{probability of transmission by same sex for 16 HIV+ states}

\item{outputs}{matrix of dim c(n.gp,2) with row of each group. second column for PrEP}

\item{ind}{index for groups of people to be mixed}

\item{ind.low}{index for groups of people at low risk of HIV transmission}

\item{ass.e}{Assortativity coefficient}

\item{n}{Number of contacts}

\item{uC}{condom use matrix}

\item{indM}{indicator of male partner}

\item{indF}{indicator of female partner}
}
\value{
Sufficient contact rate (sum of transmission via needle sharing, opposite and sex same contacts)

Complete mixing matrix
}
\description{
\code{FoI} combines parameters related to HIV transmission via needle sharing, heterosexual sex and homosexual sex
}
