% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-incidence.derivation.R
\name{incidence.derivation}
\alias{incidence.derivation}
\title{Incidence Derivation}
\usage{
incidence.derivation(CITY, case, ocis)
}
\arguments{
\item{CITY}{City name}

\item{case}{Implementation scenario}

\item{ocis}{}
}
\value{
List with disaggregated incidence outcomes
}
\description{
\code{incidence.derivation} Derive incidence data for ggplot
}
