% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-intervention.model.combination.R
\name{intervention.model.combination}
\alias{intervention.model.combination}
\title{Interventions}
\usage{
intervention.model.combination(input.parameters, current.int)
}
\arguments{
\item{input.parameters}{model inputs}

\item{current.int}{selected intervention}
}
\value{
List of population (42*19) at each time step
}
\description{
\code{intervention.model.combination} Function for running the simulation model with or without interventions
}
