% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CascadeCEA-Model-0-Function-model.initial.R
\name{model.initial}
\alias{model.initial}
\title{Set initial population}
\usage{
model.initial(par, diag18)
}
\arguments{
\item{par}{\emph{FILL IN DETAILS ABOUT PAR}}

\item{diag18}{Number of HIV-diagnosed individuals in 18 groups}
}
\value{
Matrix of initial state populations
}
\description{
\code{model.initial} sets initial model state occupancy
}
