% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CascadeCEA-Model-0-Function-ode.list.OAT.R
\name{ode.list.OAT}
\alias{ode.list.OAT}
\title{ODE equations for PWID on OAT}
\usage{
ode.list.OAT(
  x,
  x.offoat,
  beta_o,
  beta_s,
  gamma,
  rho,
  ws,
  wp,
  mu_mat,
  mo,
  eta,
  psi,
  psi.p,
  theta.ai,
  theta.ad,
  phi,
  v2,
  v3,
  alpha,
  alpha.re,
  theta.1,
  theta.2,
  theta.t12,
  theta.t13,
  theta.t21,
  theta.t23,
  theta.t31,
  theta.t32,
  theta.t1O,
  theta.t2O,
  theta.t3O,
  oat.e,
  oat.q,
  rho.m
)
}
\arguments{
\item{x}{a vector containing the number of individuals on OAT x = c(S1,S2,Sp,Ia,I1,I2,I3,Iap,Ip,Da,D1,D2,D3,T1,T2,T3,O1,O2,O3)}

\item{x.offoat}{a vector containing # of individuals off OAT}

\item{rho}{the entry rate (assume the same for all groups)}

\item{mu_mat}{maturation rates (age>65)}

\item{mo}{mortality rates for 19 states}

\item{eta}{PrEP entry rate}

\item{psi}{screening rates for susceptible or infected}

\item{psi.p}{screening rates for Iap (acute HIV on PrEP)}

\item{theta.ai}{transition rate from acute states (Ia) to chronic state (CD4>=500; I1)}

\item{theta.ad}{transition rate from acute states (Da) to chronic state (CD4>=500; D1)}

\item{phi}{\% infected of people receiving ART once diagnosed}

\item{v2, v3}{symptom-based case finding rate for I2 and I3, respectively}

\item{alpha}{ART initiation rate for D1,D2, and D3}

\item{alpha.re}{ART re-initiation rate for o1,O2, and O3}

\item{theta.1}{HIV disease progression rate for those not on ART, from I1/D1 to I2/D2}

\item{theta.2}{HIV disease progression rate for those not on ART, from I2/D2 to I3/D3}

\item{oat.e}{OAT entry rate}

\item{oat.q}{OAT dropout rate}

\item{lambda}{the sufficient contact rate}

\item{lambda.p}{the sufficient contact rate for those on PrEP.}

\item{1/ws}{average duration uninfected individuals in compartment S remain identified after screening}

\item{1/wp}{average duration uninfected individuals in compartment Sp remain on PrEP}

\item{psi.pr}{entry rate for PrEP}

\item{theta.t}{transition probabilities for those on ART}

\item{theta.o}{ART dropout probability from states T1,T2,T3}
}
\value{
HIV diagnoses, incidence and deaths for PWID on OAT
}
\description{
\code{ode.list.OAT} Calculate the change of population in cells of PWID on OAT population by the ODE model.
}
