% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Model-0-Function-ode_model-Combination.R
\name{ode_model}
\alias{ode_model}
\title{Solve ODE equations}
\usage{
ode_model(t, x, vparameters)
}
\arguments{
\item{t}{Time period (months)}

\item{x}{Index of model states}

\item{vparameters}{Model parameter inputs

Modifications to model time-varying parameters for interventions:
\enumerate{
\item Testing rate (psi)
\item ART initiation (alpha)
\item ART retention (theta)
\item Immediate ART (phi)
\item ART re-initiation (alpha prime)
\item SSP: Number of syringes available (FoI)
\item OAT: Number of individuals receiving OAT (FoI)
\item PrEP: Rate of PrEP initiation (FoI)
}}
}
\description{
\code{ode_model} ADD DETAILS
}
