% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-scale.interventions.combination.R
\name{scale.interventions.combination}
\alias{scale.interventions.combination}
\title{Interventions}
\usage{
\method{scale}{interventions.combination}(current.int, tf.comb, scale.up.period = 18)
}
\arguments{
\item{current.int}{selected intervention}

\item{tf.comb}{logical vector for possible intervention combinations}

\item{scale.up.period}{time period for intervention to reach full scale (months)}
}
\value{
Value of logical vectors for intervention modeling
}
\description{
\code{scale.interventions.combination} Function for setting up intervention combinations
}
