% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-set.scale.race.gender.R
\name{set.scale.race.gender}
\alias{set.scale.race.gender}
\title{Interventions}
\usage{
set.scale.race.gender(this.int, this.scale, these.states, otp = FALSE)
}
\arguments{
\item{this.int}{selected intervention}

\item{this.scale}{stratified scale}

\item{these.states}{affected model states}

\item{otp}{logical (default = FALSE)}
}
\value{
Value of logical vectors for intervention modeling
}
\description{
\code{set.scale.race.gender} Function for applying race/gender-specific intervention scale estimates
}
