% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-time.period.infections.R
\name{time.period.infections}
\alias{time.period.infections}
\title{Interventions}
\usage{
\method{time}{period.infections}(simul.in, start.period, end.period, sust.period)
}
\arguments{
\item{simul.in}{DESCRIPTION}

\item{start.period}{starting time point (months from initialization)}

\item{end.period}{ending time point (months from initialization)}
}
\value{
List with HIV infections outputs
}
\description{
\code{time.period.infections} Calculate HIV infections for different time periods
}
