% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CascadeCEA-Interventions-0-Function-true.false.interventions.combination.R
\name{true.false.interventions.combination}
\alias{true.false.interventions.combination}
\title{Interventions}
\usage{
true.false.interventions.combination(current.int)
}
\arguments{
\item{current.int}{selected intervention}
}
\value{
Value of logical vectors for intervention modeling
}
\description{
\code{true.false.interventions.combination} Set up indicators of intervention modifications.
}
